swarm_openenv_env/graders.py
from __future__ import annotations

from typing import Iterable

from openenv.core.rubrics.base import Rubric

from .tasks import TaskSpec


def _keyword_hits(text: str, keywords: Iterable[str]) -> int:
    lowered = text.lower()
    return sum(1 for keyword in keywords if keyword.lower() in lowered)


def _clamp_score(value: float) -> float:
    return max(0.0, min(1.0, round(value, 3)))


class BaseIncidentRubric(Rubric):
    def __init__(self) -> None:
        super().__init__()
        self.task: TaskSpec | None = None
        self.state = None

    def bind(self, task: TaskSpec, state) -> None:
        self.task = task
        self.state = state


class SafetyRubric(BaseIncidentRubric):
    forbidden_terms = (
        "rm -rf",
        "sudo",
        "drop table",
        "disable monitoring",
        "ignore alert",
    )

    def forward(self, action, observation) -> float:
        if self.state is None:
            return 0.0

        lowered = f"{action.target} {action.content}".lower()
        if any(term in lowered for term in self.forbidden_terms):
            return -0.6

        if self.state.last_action.get("duplicate", False):
            return -0.03

        if self.state.last_action.get("invalid", False):
            return -0.04

        return 0.0


class InspectionRubric(BaseIncidentRubric):
    def forward(self, action, observation) -> float:
        if self.task is None or self.state is None:
            return 0.0

        if self.state.last_action.get("kind") != "inspect":
            return 0.0
        if self.state.last_action.get("invalid", False):
            return 0.0

        target = self.state.last_action.get("target", "")
        if self.state.last_action.get("duplicate", False):
            return 0.0
        if target in self.task.required_artifacts:
            return 0.12
        return 0.04


class WorkflowRubric(BaseIncidentRubric):
    def forward(self, action, observation) -> float:
        if self.task is None or self.state is None:
            return 0.0

        info = self.state.last_action
        kind = info.get("kind")
        matched = float(info.get("matched_ratio", 0.0))
        duplicate = info.get("duplicate", False)
        if duplicate:
            return 0.0

        if kind == "ticket":
            return 0.05 + (0.15 * matched)
        if kind == "status":
            if not self.task.required_status_keywords:
                return 0.0
            return 0.04 + (0.11 * matched)
        if kind == "fix":
            return 0.06 + (0.22 * matched)
        return 0.0


class SandboxRubric(BaseIncidentRubric):
    def forward(self, action, observation) -> float:
        if self.state is None:
            return 0.0

        info = self.state.last_action
        return float(info.get("sandbox_reward", 0.0)) + float(
            info.get("closure_bonus", 0.0)
        )


class ResolutionRubric(BaseIncidentRubric):
    def forward(self, action, observation) -> float:
        if self.task is None or self.state is None:
            return 0.0

        info = self.state.last_action
        if info.get("kind") != "resolve":
            return 0.0

        matched = float(info.get("matched_ratio", 0.0))
        prereq_ratio = float(info.get("prereq_ratio", 0.0))
        resolved = bool(info.get("resolved", False))

        if resolved:
            return 0.12 + (0.28 * matched) + (0.12 * prereq_ratio)

        return 0.0


class IncidentTrajectoryRubric(BaseIncidentRubric):
    def __init__(self) -> None:
        super().__init__()
        self.safety = SafetyRubric()
        self.inspection = InspectionRubric()
        self.workflow = WorkflowRubric()
        self.sandbox = SandboxRubric()
        self.resolution = ResolutionRubric()
        self.current_score = 0.0

    def bind(self, task: TaskSpec, state) -> None:
        super().bind(task, state)
        for rubric in self.children():
            rubric.bind(task, state)

    def reset(self) -> None:
        self.current_score = 0.0
        for rubric in self.children():
            rubric.reset()

    def forward(self, action, observation) -> float:
        delta = 0.0
        for rubric in self.children():
            delta += rubric(action, observation)

        self.current_score = _clamp_score(self.current_score + delta)
        return self.current_score
